+++ src/interfaces/ISimpleMove.ts
@@ -1,0 +1,8 @@
+export interface ISimpleMove {
+  player: string;
+  after: string;
+  before: string;
+  piece: string;
+  from: string;
+  to: string;
+}
